Laboratory 7

+++ app/Database/Migrations/20251017031019_AddRoleToUsers.php
@@ -6,14 +6,20 @@ class AddRoleToUsers extends Migration
 {
     public function up()
     {
-        $fields = [
-            'role' => ['type' => "VARCHAR", 'constraint' => 50, 'null' => false, 'default' => 'student'],
-        ];
-        $this->forge->addColumn('users', $fields);
+        $db = \Config\Database::connect();
+        if (!$db->fieldExists('role', 'users')) {
+            $fields = [
+                'role' => ['type' => "VARCHAR", 'constraint' => 50, 'null' => false, 'default' => 'student'],
+            ];
+            $this->forge->addColumn('users', $fields);
+        }
     }
 
     public function down()
     {
-        $this->forge->dropColumn('users', 'role');
+        $db = \Config\Database::connect();
+        if ($db->fieldExists('role', 'users')) {
+            $this->forge->dropColumn('users', 'role');
+        }
     }
 }
